Moved verification completely to separate module, gated by non-secure feature
